+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Change History
 Version 0.8.1 (Pending)
 - Changed login/logout pages to use site template/theme
+- Added attendee edit with override from another user that has edit right
 
 Version 0.8.0 (8/18/2019)
 - Tightened up formatting on at-con registration forms